requirements.txt
pyserial>=3.3
RPi.GPIO==0.6.3
